+++ src/Common/SearchHeader/Header/HomeHeader.module.css
@@ -1,85 +1,85 @@
 /* hyein's style */
 
-header {
-   position: absolute;
-  top: 0;
-  left: 50%;
- transform: translateX(-50%);
-  display: flex;
-  align-items: center;
-  width: 21.4375rem;
+/*header {*/
+/*   position: absolute;*/
+/*  top: 0;*/
+/*  left: 50%;*/
+/* transform: translateX(-50%);*/
+/*  display: flex;*/
+/*  align-items: center;*/
+/*  width: 21.4375rem;*/
 
-  justify-content: center;
-  color: white;
-  gap: 6.6875rem;
-  font-weight: bold;
-  background: var(--main-color, #38ACFF);
-  flex-shrink: 0;
-  padding: 1rem;
+/*  justify-content: center;*/
+/*  color: white;*/
+/*  gap: 6.6875rem;*/
+/*  font-weight: bold; */
+/*  background: var(--main-color, #38ACFF);*/
+/*  flex-shrink: 0;*/
+/*  padding: 1rem;*/
 
-  /* display: flex;
-width: 23.4375rem;
-height: 3.5rem;
+/*  !* display: flex;*/
+/*width: 23.4375rem;*/
+/*height: 3.5rem;*/
 
-justify-content: center;
-align-items: flex-start;
+/*justify-content: center;*/
+/*align-items: flex-start;*/
 
-   */
+/*   *!*/
 
-}
-
-.logo {
-  width: 5.125rem;
-  height: 1.375rem;
-  flex-shrink: 0;}
-
-
-
-.icon {
-  display: flex;
-  width: 24px;
-  height: 24px;
-  margin-left:1;
-  margin-right:1;
-}
-
-/* ============================================================== */
-/* dongbeom's style */
-/*.header {*/
-/*    position: absolute;*/
-
-/*    display: flex;*/
-/*    width: 100%;*/
-/*    !* padding을 준 만큼 width, height 길이 제거 *!*/
-/*    max-width: 21.4375rem;*/
-/*    height: 1.5rem;*/
-/*    padding: 1rem;*/
-/*    justify-content: center;*/
-/*    align-items: flex-start;*/
-/*    gap: 6.6875rem;*/
-/*    flex-shrink: 0;*/
-
-/*    background: var(--main-color, #38ACFF);*/
 /*}*/
 
 /*.logo {*/
-/*    display: flex;*/
-/*    width: 1.5rem;*/
-/*    height: 1.5rem;*/
-/*    justify-content: center;*/
-/*    align-items: center;*/
-/*    flex-shrink: 0;*/
-/*}*/
+/*  width: 5.125rem;*/
+/*  height: 1.375rem;*/
+/*  flex-shrink: 0;}*/
+
+
 
 /*.icon {*/
-/*    display: flex;*/
-/*    width: 1.5rem;*/
-/*    height: 1.5rem;*/
-/*    justify-content: center;*/
-/*    align-items: center;*/
-/*    flex-shrink: 0;*/
+/*  display: flex;*/
+/*  width: 24px;*/
+/*  height: 24px;*/
+/*  margin-left:1;*/
+/*  margin-right:1;*/
 /*}*/
 
-/*.icon, .logo:hover {*/
-/*    cursor: pointer;*/
-/*}*/
+/* ============================================================== */
+/* dongbeom's style */
+.header {
+    position: absolute;
+
+    display: flex;
+    width: 100%;
+    /* padding을 준 만큼 width, height 길이 제거 */
+    max-width: 21.4375rem;
+    height: 1.5rem;
+    padding: 1rem;
+    justify-content: center;
+    align-items: flex-start;
+    gap: 6.6875rem;
+    flex-shrink: 0;
+
+    background: var(--main-color, #38ACFF);
+}
+
+.logo {
+    display: flex;
+    width: 1.5rem;
+    height: 1.5rem;
+    justify-content: center;
+    align-items: center;
+    flex-shrink: 0;
+}
+
+.icon {
+    display: flex;
+    width: 1.5rem;
+    height: 1.5rem;
+    justify-content: center;
+    align-items: center;
+    flex-shrink: 0;
+}
+
+.icon, .logo:hover {
+    cursor: pointer;
+}
