+++ src/Common/SearchHeader/Input/InputField.module.css
@@ -1,110 +1,65 @@
 /* hyein's style */
 
-.Wrapper {
-  display: flex;
-
-  width: 21.4375rem;
-   /* 최대 너비를 21.4375rem로 설정 */
-  padding: 1rem;
-  justify-content: center;
-  align-items: flex-start;
-  flex-shrink: 0;
-  background: var(--primary, #38ACFF);
-  margin: 2.8rem auto 0;
-}
-
-.SearchContainer {
-  width: 100%;
-  display: flex;
-  align-items: center;
-  height: 2.875rem; /* 46px */
-  justify-content: center;
-  margin: auto 0;
-  flex-shrink: 0;
-  border-radius: 1.25rem; /* 20px */
-  position: relative;
-}
-
-.SearchBar::placeholder {
-  color: var(--text-light);
-}
-
-.SearchBar {
-  flex: 1;
-  height: 100%;
-  padding: 0 1.25rem;
-  border: none;
-  outline: none;
-  border-radius: 20px;
-  font-size: 1rem;
-  font-weight: 500;
-  line-height: 1.6rem; /* 25.6px */
-  letter-spacing: 0;
-  text-align:left;
-}
-
-.SearchButton {
-  gap: 6.8125rem;
-  align-items: center;
-  justify-content: center;
-  width: 24px; /* 아이콘 크기 지정 */
-  height: 24px; /* 아이콘 크기 지정 */
-  position: absolute; /* SearchContainer를 기준으로 위치 조정 */
-  right: 1rem; /* 오른쪽 여백 추가 */
-  background: none; /* 버튼의 배경 제거 */
-  border: none; /* 버튼의 테두리 제거 */
-  cursor: pointer; /* 클릭 가능한 포인터 설정 */
-}
-
-.SearchButton img {
-  width: 24px; /* 이미지 크기 지정 */
-  height: 24px; /* 이미지 크기 지정 */
-
-}
-
-.SearchBar:focus {
-  border-color: var(--main-color, #38ACFF);
-  outline: none;
-}
-
-/* =================================================================== */
-/* dongbeom's style */
-
 /*.Wrapper {*/
-/*  position: fixed;*/
-
 /*  display: flex;*/
-/*  width: 100%; !* 너비를 100%로 설정 *!*/
-/*  max-width: 21.4375rem; !* 최대 너비를 21.4375rem로 설정 *!*/
-/*  !* height:4.875rem; *!*/
-/*  padding: 0rem 1rem 1rem 1rem;*/
+/* */
+/*  width: 21.4375rem;*/
+/*   !* 최대 너비를 21.4375rem로 설정 *!*/
+/*  padding: 1rem;*/
 /*  justify-content: center;*/
 /*  align-items: flex-start;*/
-/*  !* gap: 6.8125rem; *!*/
 /*  flex-shrink: 0;*/
-/*  background: var(--main-color, #38ACFF);*/
-/*  margin: 3.5rem auto 0;*/
-/*  !* margin-left: -0.4rem ; *!*/
-
+/*  background: var(--primary, #38ACFF);*/
+/*  margin: 2.8rem auto 0;*/
 /*}*/
+
 /*.SearchContainer {*/
 /*  width: 100%;*/
 /*  display: flex;*/
 /*  align-items: center;*/
-/*  height: 2.875rem;   !* 46px *!*/
-
+/*  height: 2.875rem; !* 46px *!*/
 /*  justify-content: center;*/
 /*  margin: auto 0;*/
 /*  flex-shrink: 0;*/
-/*  border-radius: 1.25rem;  !* 20px *!*/
-
+/*  border-radius: 1.25rem; !* 20px *!*/
 /*  position: relative;*/
-/*  !* 아이콘의 위치 조정을 위해 position 설정 *!*/
-
 /*}*/
 
 /*.SearchBar::placeholder {*/
-/*  color: var(--text-light, #C0C0C0);*/
+/*  color: var(--text-light);*/
+/*}*/
+
+/*.SearchBar {*/
+/*  flex: 1;*/
+/*  height: 100%;*/
+/*  padding: 0 1.25rem;*/
+/*  border: none;*/
+/*  outline: none;*/
+/*  border-radius: 20px;*/
+/*  font-size: 1rem;*/
+/*  font-weight: 500;*/
+/*  line-height: 1.6rem; !* 25.6px *!*/
+/*  letter-spacing: 0;*/
+/*  text-align:left;*/
+/*}*/
+
+/*.SearchButton {*/
+/*  gap: 6.8125rem;*/
+/*  align-items: center;*/
+/*  justify-content: center;*/
+/*  width: 24px; !* 아이콘 크기 지정 *!*/
+/*  height: 24px; !* 아이콘 크기 지정 *!*/
+/*  position: absolute; !* SearchContainer를 기준으로 위치 조정 *!*/
+/*  right: 1rem; !* 오른쪽 여백 추가 *!*/
+/*  background: none; !* 버튼의 배경 제거 *!*/
+/*  border: none; !* 버튼의 테두리 제거 *!*/
+/*  cursor: pointer; !* 클릭 가능한 포인터 설정 *!*/
+/*}*/
+
+/*.SearchButton img {*/
+/*  width: 24px; !* 이미지 크기 지정 *!*/
+/*  height: 24px; !* 이미지 크기 지정 *!*/
+/* */
 /*}*/
 
 /*.SearchBar:focus {*/
@@ -112,40 +67,85 @@
 /*  outline: none;*/
 /*}*/
 
-/*.SearchBar {*/
-/*  flex: 1;*/
-/*  !* 남은 공간을 모두 차지하도록 설정 *!*/
-/*  height: 100%;*/
-/*  padding: 0 1.25rem;  !* 20px *!*/
-/*  !* 좌우 padding을 추가하여 텍스트와 경계 사이에 여백을 만듦 *!*/
-/*  border: none;*/
-/*  outline: none;*/
-/*  !* 기본 포커스 스타일 제거 *!*/
-/*  border-radius: 1.25rem;*/
-/*  background: var(--white, #FFFFFF);*/
+/* =================================================================== */
+/* dongbeom's style */
 
-/*  !* Body2 스타일 *!*/
-/*  font-size: 1rem;  !* 16px *!*/
-/*  font-weight: 500;*/
-/*  line-height: 1.6rem;  !* 25.6px *!*/
-/*  letter-spacing: 0;*/
-/*}*/
+.Wrapper {
+  position: fixed;
 
-/*!* Wrapper에 padding 수정하고 그에 맞춰 SearchButton 수정했습니다.(padding, width, height) *!*/
+  display: flex;
+  width: 100%; /* 너비를 100%로 설정 */
+  max-width: 21.4375rem; /* 최대 너비를 21.4375rem로 설정 */
+  /* height:4.875rem; */
+  padding: 0rem 1rem 1rem 1rem;
+  justify-content: center;
+  align-items: flex-start;
+  /* gap: 6.8125rem; */
+  flex-shrink: 0;
+  background: var(--main-color, #38ACFF);
+  margin: 3.5rem auto 0;
+  /* margin-left: -0.4rem ; */
 
-/*.SearchButton {*/
-/*  align-items: center;*/
-/*  justify-content: center;*/
-/*  position: absolute; !* SearchContainer를 기준으로 위치 조정 *!*/
-/*  right: 1rem; !* 오른쪽 여백 추가 *!*/
-/*  background: none; !* 버튼의 배경 제거 *!*/
-/*  border: none; !* 버튼의 테두리 제거 *!*/
-/*  cursor: pointer; !* 클릭 가능한 포인터 설정 *!*/
+}
+.SearchContainer {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  height: 2.875rem;   /* 46px */
 
-/*  padding: 0rem;*/
-/*}*/
+  justify-content: center;
+  margin: auto 0;
+  flex-shrink: 0;
+  border-radius: 1.25rem;  /* 20px */
 
-/*.SearchButton img {*/
-/*  width: 1.5rem; !* 이미지 크기 지정 *!*/
-/*  height: 1.5rem; !* 이미지 크기 지정 *!*/
-/*}*/
+  position: relative;
+  /* 아이콘의 위치 조정을 위해 position 설정 */
+
+}
+
+.SearchBar::placeholder {
+  color: var(--text-light, #C0C0C0);
+}
+
+.SearchBar:focus {
+  border-color: var(--main-color, #38ACFF);
+  outline: none;
+}
+
+.SearchBar {
+  flex: 1;
+  /* 남은 공간을 모두 차지하도록 설정 */
+  height: 100%;
+  padding: 0 1.25rem;  /* 20px */
+  /* 좌우 padding을 추가하여 텍스트와 경계 사이에 여백을 만듦 */
+  border: none;
+  outline: none;
+  /* 기본 포커스 스타일 제거 */
+  border-radius: 1.25rem;
+  background: var(--white, #FFFFFF);
+
+  /* Body2 스타일 */
+  font-size: 1rem;  /* 16px */
+  font-weight: 500;
+  line-height: 1.6rem;  /* 25.6px */
+  letter-spacing: 0;
+}
+
+/* Wrapper에 padding 수정하고 그에 맞춰 SearchButton 수정했습니다.(padding, width, height) */
+
+.SearchButton {
+  align-items: center;
+  justify-content: center;
+  position: absolute; /* SearchContainer를 기준으로 위치 조정 */
+  right: 1rem; /* 오른쪽 여백 추가 */
+  background: none; /* 버튼의 배경 제거 */
+  border: none; /* 버튼의 테두리 제거 */
+  cursor: pointer; /* 클릭 가능한 포인터 설정 */
+
+  padding: 0rem;
+}
+
+.SearchButton img {
+  width: 1.5rem; /* 이미지 크기 지정 */
+  height: 1.5rem; /* 이미지 크기 지정 */
+}
