+++ src/components/HambergerMenu/SubMenu.jsx
@@ -2,10 +2,10 @@ import { useEffect, useState } from 'react'
 import { useLocation, useNavigate } from 'react-router-dom';
 import axios from 'axios'
 import styles from './SubMenu.module.css'
-import layout from '../Common/TestLayout.module.css'
-import img from '../assets/img/menu/book_icon_black.svg'
-import img2 from '../assets/img/menu/library.svg'
-import logoutimg from '../assets/img/menu/logout.svg'
+import layout from '../../Common/TestLayout.module.css'
+import img from '../../assets/img/menu/book_icon_black.svg'
+import img2 from '../../assets/img/menu/library.svg'
+import logoutimg from '../../assets/img/menu/logout.svg'
 
 export default function SubMenu() {
     const [email, setEmail] = useState("로그인이 필요합니다");
