+++ src/pages/LibraryLocation/Header/Header.module.css
@@ -1,6 +1,6 @@
 @font-face {
     font-family : 'Pretendard';
-    src: url("../../assets/font/PretendardVariable.ttf");
+    src: url("../../../assets/font/PretendardVariable.ttf");
 }
 
 .layout {
